+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 
 ### Bug Fixes
 
-* announce single task completion only 52f7053
+* debounce codex-helper single task completion event 10cd9ac
 * include download urls in userscript metadata (#2) 92a467f, closes #2